+++ tests/jobflow/test_customizers.py
@@ -5,7 +5,7 @@ import pytest
 jf = pytest.importorskip("jobflow")
 from pathlib import Path
 
-from quacc import SETTINGS, job, redecorate, strip_decorator
+from quacc import job, redecorate, strip_decorator
 
 
 def test_strip_decorators():
@@ -22,6 +22,7 @@ def test_change_settings_redecorate_job(tmp_path_factory):
 
     @job
     def write_file_job(name="job.txt"):
+        from quacc import SETTINGS
         with open(Path(SETTINGS.RESULTS_DIR, name), "w") as f:
             f.write("test file")
